+++ classes/terrain.cpp
@@ -6,11 +6,10 @@
 
 
 Terrain::Terrain(const TERRAIN_TYPE terrain_type,
-                 const uint texture_index, int position[2]): Object(texture_index, position) {
+                 const uint texture_index, float position[2]): Object(texture_index, position) {
     tags[0] = TERRAIN;
     tags[1] = OWNER_WORLD;
     tags[2] = INDESTRUCTIBLE;
     type = terrain_type;
-
     occupied = false;
 }
